% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_continuous.R
\name{Sim_continuous}
\alias{Sim_continuous}
\title{Simulation of continuous outcome in the paper}
\usage{
Sim_continuous(n, beta1, beta2, q, q1, q2, q3, c1, var_g, mu, mu1, sigma_y)
}
\arguments{
\item{n}{number of total samples (the first n/3 samples belong to cluster 1, next n/3 belongs to cluster 2 and the last n/3 belong to cluster 3)}

\item{beta1}{coefficent of first covariate}

\item{beta2}{coefficent of second covariate}

\item{q}{number of true informative genes}

\item{q1}{number of genes for random clusters}

\item{q2}{number of genes correlated with covariates}

\item{q3}{number of noise genes}

\item{c1}{effect size of outcome separation}

\item{var_g}{control the correlation of the q3 genes with covariates}

\item{mu}{effect size of gene separation}

\item{mu1}{effect size of random clusters}

\item{sigma_y}{standard deviation for outcome}
}
\value{
result_list: A list with three components.
\itemize{
\item{x: }{simulated covariate matrix}
\item{y: }{simulated outcome}
\item{G: }{simulated gene expression data}
}
}
\description{
Simulation of continuous outcome in the paper
}
\examples{
\dontrun{
Data<-Sim_surv(n=99,beta1 = 0.5,beta2 = 0.5,q = 50,q1 = 300,q2 = 50,q3 = 1600,c1 = 2,var_g = 2,mu = 1.5,mu1 = 1.7,sigma_y = 0.5,censor = 100)
Data<-Sim_continuous(n=99,beta1 = 1,beta2 = 1,q = 50,q1 = 50,q2 = 50,q3 = 1850,c1 = 3,var_g = 2,mu = 1.5,mu1 = 1.7,sigma_y = 1)
G<-Data$G
X<-Data$x
Y<-Data$y
}
}
